+++ CHANGES.rst
@@ -5,7 +5,8 @@ Changelog
 1.0.5 (unreleased)
 ------------------
 
-- Nothing changed yet.
+- Added revoke-user-access page to remove a user from its groups and revoke its roles.
+  [odelaere]
 
 
 1.0.4 (2020-10-08)